+++ src/yule_orchestrator/core/__init__.py
@@ -1,0 +1,9 @@
+from .context_loader import ContextError, ContextDocument, LoadedContext, load_agent_context, render_context
+
+__all__ = [
+    "ContextError",
+    "ContextDocument",
+    "LoadedContext",
+    "load_agent_context",
+    "render_context",
+]
