+++ src/yule_orchestrator/diagnostics/__init__.py
@@ -1,0 +1,3 @@
+from .doctor import DoctorCheck, doctor_exit_code, render_doctor_report, run_doctor
+
+__all__ = ["DoctorCheck", "doctor_exit_code", "render_doctor_report", "run_doctor"]
