+++ src/yule_orchestrator/integrations/github/__init__.py
@@ -1,0 +1,3 @@
+from .issues import GitHubIssue, GitHubIssueError, list_open_issues, render_open_issues
+
+__all__ = ["GitHubIssue", "GitHubIssueError", "list_open_issues", "render_open_issues"]
